% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-documentation.R
\docType{data}
\name{GSE10797}
\alias{GSE10797}
\title{SummarizedExperiment object GSE10797 containing data for genes and samples.}
\format{
A SummarizedExperiment object with:
\describe{
\item{assays}{a matrix of counts (genes x samples)}
\item{rowData}{feature data (data related to genes)}
\item{colData}{sample metadata (data related to samples)}
}
}
\source{
Generated internally for RPracticePackage
}
\usage{
GSE10797
}
\description{
SummarizedExperiment object GSE10797 containing data for genes and samples.
}
\keyword{datasets}
