% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-documentation.R
\docType{data}
\name{GSE41197}
\alias{GSE41197}
\title{SummarizedExperiment object GSE41197 containing data for genes and samples.}
\format{
A SummarizedExperiment object with:
\describe{
\item{assays}{a matrix of counts (genes x samples)}
\item{rowData}{feature data (data related to genes)}
\item{colData}{sample metadata (data related to samples)}
}
}
\source{
Generated internally for RPracticePackage
}
\usage{
GSE41197
}
\description{
SummarizedExperiment object GSE41197 containing data for genes and samples.
}
\keyword{datasets}
